Same entry bugfix
Bug description
If there are two or more entries with the same
content, there is a problem while toggling or
deleting. Specifically, always the top most
entry is toggled/deleted.
Solution:
Mark the created li elements also with an id,
same as the one coming from the server/db.
Change the way of toggling/deleting, based now
on the id instead of the lexical content of the
entries.